[MOV] website: move MockRequest to http_routing.tests.common
`MockRequest` is used for testing. When running odoo, we should not have
this class nor import unittest framework.
The class should not be in base because it uses some things defined in
modules (http_routing).
