+++ addons/test_website/tests/test_fuzzy.py
@@ -5,7 +5,7 @@ import logging
 import psycopg2
 
 from odoo.addons.website.controllers.main import Website
-from odoo.addons.website.tools import MockRequest
+from odoo.addons.http_routing.tests.common import MockRequest
 import odoo.tests
 from odoo.tests.common import TransactionCase
