+++ addons/website/tests/test_fuzzy.py
@@ -1,14 +1,17 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 import logging
-from lxml import etree
 import re
+
+from lxml import etree
 from markupsafe import Markup
 
-from odoo.addons.website.controllers.main import Website
-from odoo.addons.website.tools import distance, MockRequest
 import odoo.tests
 from odoo.tests.common import TransactionCase
+
+from odoo.addons.website.controllers.main import Website
+from odoo.addons.http_routing.tests.common import MockRequest
+from odoo.addons.website.tools import distance
 
 _logger = logging.getLogger(__name__)
