+++ addons/website/tools.py
@@ -1,117 +1,11 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
-import contextlib
 import re
+
 import werkzeug.urls
 from lxml import etree
-from unittest.mock import Mock, MagicMock, patch
 
-from werkzeug.exceptions import NotFound
-from werkzeug.test import EnvironBuilder
+from odoo.tools.misc import hmac
 
-import odoo.http
-from odoo.tools.misc import hmac, DotDict, frozendict
-
-HOST = '127.0.0.1'
-
-
-@contextlib.contextmanager
-def MockRequest(
-    env, *, path='/mockrequest', routing=True, multilang=True,
-    context=frozendict(), cookies=frozendict(), country_code=None,
-    website=None, remote_addr=HOST, environ_base=None, url_root=None,
-):
-    # TODO move MockRequest to a package in addons/web/tests
-    from odoo.tests.common import HttpCase  # noqa: PLC0415
-    lang_code = context.get('lang', env.context.get('lang', 'en_US'))
-    env = env(context=dict(context, lang=lang_code))
-    if HttpCase.http_port():
-        base_url = HttpCase.base_url()
-    else:
-        base_url = f"http://{HOST}:{odoo.tools.config['http_port']}"
-    request = Mock(
-        # request
-        httprequest=Mock(
-            host='localhost',
-            path=path,
-            app=odoo.http.root,
-            environ=dict(
-                EnvironBuilder(
-                    path=path,
-                    base_url=base_url,
-                    environ_base=environ_base,
-                ).get_environ(),
-                REMOTE_ADDR=remote_addr,
-            ),
-            cookies=cookies,
-            referrer='',
-            remote_addr=remote_addr,
-            url_root=url_root,
-            args=[],
-        ),
-        type='http',
-        future_response=odoo.http.FutureResponse(),
-        params={},
-        redirect=env['ir.http']._redirect,
-        session=DotDict(
-            odoo.http.get_default_session(),
-            context={'lang': ''},
-            force_website_id=website and website.id,
-        ),
-        geoip=odoo.http.GeoIP('127.0.0.1'),
-        db=env.registry.db_name,
-        env=env,
-        registry=env.registry,
-        cr=env.cr,
-        uid=env.uid,
-        context=env.context,
-        cookies=cookies,
-        lang=env['res.lang']._get_data(code=lang_code),
-        website=website,
-        render=lambda *a, **kw: '<MockResponse>',
-    )
-    if url_root is not None:
-        request.httprequest.url = werkzeug.urls.url_join(url_root, path)
-    if website:
-        request.website_routing = website.id
-    if country_code:
-        try:
-            request.geoip._city_record = odoo.http.geoip2.models.City(['en'], country={'iso_code': country_code})
-        except TypeError:
-            request.geoip._city_record = odoo.http.geoip2.models.City({'country': {'iso_code': country_code}})
-
-    # The following code mocks match() to return a fake rule with a fake
-    # 'routing' attribute (routing=True) or to raise a NotFound
-    # exception (routing=False).
-    #
-    #   router = odoo.http.root.get_db_router()
-    #   rule, args = router.bind(...).match(path)
-    #   # arg routing is True => rule.endpoint.routing == {...}
-    #   # arg routing is False => NotFound exception
-    router = MagicMock()
-    match = router.return_value.bind.return_value.match
-    if routing:
-        match.return_value[0].routing = {
-            'type': 'http',
-            'website': True,
-            'multilang': multilang
-        }
-    else:
-        match.side_effect = NotFound
-
-    def update_context(**overrides):
-        request.env = request.env(context=dict(request.context, **overrides))
-        request.context = request.env.context
-
-    request.update_context = update_context
-
-    with contextlib.ExitStack() as s:
-        odoo.http._request_stack.push(request)
-        s.callback(odoo.http._request_stack.pop)
-        s.enter_context(patch('odoo.http.root.get_db_router', router))
-
-        yield request
-
-# Fuzzy matching tools
 
 def distance(s1="", s2="", limit=4):
     """
@@ -150,6 +44,7 @@ def distance(s1="", s2="", limit=4):
         p, d = d, p
     return p[l1] if p[l1] <= limit else -1
 
+
 def similarity_score(s1, s2):
     """
     Computes a score that describes how much two strings are matching.
@@ -168,6 +63,7 @@ def similarity_score(s1, s2):
     score -= dist / len(s1)
     score -= len(set1.symmetric_difference(s2)) / (len(s1) + len(s2))
     return score
+
 
 def text_from_html(html_fragment, collapse_whitespace=False):
     """
@@ -193,8 +89,9 @@ def text_from_html(html_fragment, collapse_whitespace=False):
 
     content = ' '.join(tree.itertext())
     if collapse_whitespace:
-        content = re.sub('\\s+', ' ', content).strip()
+        content = re.sub(r'\s+', ' ', content).strip()
     return content
+
 
 def get_base_domain(url, strip_www=False):
     """
