+++ addons/website_sale/tests/common.py
@@ -7,7 +7,7 @@ from odoo.tools import lazy
 
 from odoo.addons.delivery.tests.common import DeliveryCommon
 from odoo.addons.product.tests.common import ProductCommon
-from odoo.addons.website.tools import MockRequest as websiteMockRequest
+from odoo.addons.http_routing.tests.common import MockRequest as websiteMockRequest
 from odoo.addons.website_sale.models.website import (
     CART_SESSION_CACHE_KEY,
     FISCAL_POSITION_SESSION_CACHE_KEY,
