+++ addons/website_sale/tests/test_website_sale_snippets.py
@@ -4,7 +4,7 @@ import logging
 
 from odoo.tests import HttpCase, tagged
 
-from odoo.addons.website.tools import MockRequest
+from odoo.addons.http_routing.tests.common import MockRequest
 
 
 _logger = logging.getLogger(__name__)
